Remove 'buildUsagesItems' from `EditorComponent`
* `EditorConst` now public
* `gotoElement` now public final
* few refactoring
